change main.go, /api --> cmd/api/ and cmd/main.go. Add GET: "/ping" endpoint
