+++ cmd/main.go
@@ -4,7 +4,7 @@ import (
 	"log"
 	"os"
 
-	"github.com/crisnlopez/social-media-bkend/api"
+	"github.com/crisnlopez/social-media-bkend/cmd/api"
 )
 
 const defaultPort = "8080"
@@ -17,5 +17,5 @@ func main() {
 		port = defaultPort
 	}
 
-	api.Start(port)
+  api.Start(port)
 }
